resolves #507
